requirement.txt
langchain==0.0.350
transformers==4.36.0
torch==2.1.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
pytest==7.4.0
python-magic==0.4.27
tqdm==4.65.0